fix NavDrawer view
